Double naar int aangepast

+++ Tablr/src/sql/GreaterThanExpression.java
@@ -5,16 +5,16 @@ import java.util.ArrayList;
 import domain.Table;
 
 public class GreaterThanExpression extends Expression<Boolean> {
-	public GreaterThanExpression(Expression<Double> expression1, Expression<Double> expression2) {
+	public GreaterThanExpression(Expression<Integer> expression1, Expression<Integer> expression2) {
 		this.expression1 = expression1;
 		this.expression2 = expression2;
 	}
 	
-	private Expression<Double> expression1;
-	private Expression<Double> expression2;
+	private Expression<Integer> expression1;
+	private Expression<Integer> expression2;
 	
 	public Boolean eval(ArrayList<Table> tables, int rowNb) {
-		return ((Double)expression1.eval(tables, rowNb) > (Double)expression2.eval(tables, rowNb));
+		return ((int)expression1.eval(tables, rowNb) > (int)expression2.eval(tables, rowNb));
 	}
 	
 	public String toString() {
